moved typeout code from debugger.c to typeout.c
