+++ src/typeout.h
@@ -1,0 +1,16 @@
+typedef void (typeoutfunc)(uint64_t);
+
+extern typeoutfunc *mdquot;
+extern typeoutfunc *mnmsgn;
+extern typeoutfunc *mdolla;
+extern typeoutfunc *mperce;
+extern typeoutfunc *mamper;
+extern typeoutfunc *mprime;
+extern typeoutfunc *tch;
+extern typeoutfunc *sch;
+
+void tma(uint64_t value);
+void tmc(uint64_t value);
+void tmch(uint64_t value);
+void tmf(uint64_t value);
+void tmh(uint64_t value);
